+++ telegram-bot/src/main/kotlin/eu/vendeli/tgbot/types/Document.kt
@@ -6,5 +6,5 @@ data class Document(
     val thumb: PhotoSize? = null,
     val fileName: String? = null,
     val mimeType: String? = null,
-    val fileSize: Int?,
+    val fileSize: Long?,
 )
